requirements.txt
Django==4.2.7
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
reportlab==4.0.7
qrcode[pil]==7.4.2
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
daphne==4.2.1
djangorestframework==3.14.0  # REST API for print clients
python-decouple==3.8         # Environment variable management

# Security packages
django-ratelimit==4.1.0      # Rate limiting for views
django-axes==6.1.1           # Failed login attempt tracking
django-cors-headers==4.3.1   # CORS headers management
bleach==6.1.0                # HTML sanitization
